Split getProfileToCertifyV1 tests to integration
